+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/pusher/pusher-platform-swift/compare/0.1.25...HEAD)
+## [Unreleased](https://github.com/pusher/pusher-platform-swift/compare/0.1.26...HEAD)
+
+## [0.1.26](https://github.com/pusher/pusher-platform-swift/compare/0.1.25...0.1.26) - 2017-07-24
+### Changed
+- Rename `App` to `Instance`
+- `Instance` and `PPBaseClient` constructor
 
 ## [0.1.25](https://github.com/pusher/pusher-platform-swift/compare/0.1.24...0.1.25) - 2017-07-19
 ### Changed